detection/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "parallels"
version = "0.1.0"
description = "Find parallel text segments across a corpus using FAISS"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "faiss-cpu>=1.7.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.60.0",
    "openpyxl>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
parallels = "parallels.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["parallels*"]
